Fix code review issues: improve precision and security

+++ src/rpc.rs
@@ -1099,11 +1099,12 @@ pub async fn sell_token(
         info!("Added close_account instruction to reclaim ~0.00203928 SOL rent from ATA {}", ata);
         
         // Add dev fee instruction (2% of sell proceeds)
+        // Use integer arithmetic to avoid floating-point precision errors
         if let Some(dev_wallet_str) = &settings.dev_fee_wallet {
             if let Ok(dev_wallet) = Pubkey::from_str(dev_wallet_str) {
                 // Calculate dev fee based on expected SOL received from sell
                 let expected_sol_lamports = (amount as f64 * current_price * 1_000_000_000.0) as u64;
-                let dev_fee_lamports = (expected_sol_lamports as f64 * (settings.dev_fee_bps as f64 / 10000.0)) as u64;
+                let dev_fee_lamports = (expected_sol_lamports * settings.dev_fee_bps) / 10000;
                 let dev_fee_instr = solana_sdk::system_instruction::transfer(&user_pubkey, &dev_wallet, dev_fee_lamports);
                 all_instrs.push(dev_fee_instr);
                 info!("Added dev fee: {:.6} SOL ({} basis points) to {}",
